requirements.txt
requests
pandas
fastapi
uvicorn[standard]
psycopg2-binary
plotly
